Conditionally render legacy sidebar
The sidebar provided additional information about a subscription. It
made several assumptions about a subscription having a single
subscription line item.

There are new designs for the subscription admin area, so in order to
fix the admin portal in the short term before the new design are
implemented I am simply moving the sidebar to a partial and rendering it
only if the Subscription object responds to the correct api